append ToCore to all errors tests
In order to retrieve all clues structured error data
in tests, we need to extract it from the error using
the clues library.

This change appends `clues.ToCore(err)` to all
variations of `assert.NoError(t, err)`.  The only
other changes are those necessary to preserve
linting, or to produce an error variable for the
ToCore call.