Make sure ShortRef is unique after munging
Take advantage of how data should be laid out in OneDrive to ensure
ShortRefs will be unique after munging. This relies on not being able to
make a folder with the same name as a file in the same directory. By
arranging all the display names before the item ID, we ensure that.

Add a test to make sure that the generated ShortRef isn't equal to the
RepoRef item name followed by the display name of the file.