scopes typing rename, filter -> info
Renames a lot of the names in scopes that contain
the word "filter" with the word "info".  The basis
here is to reduce confusion in the libraries.  Filter
as a term is overloaded.  It currently means: 1/
the set of scopes that comparisons _must_ pass
to succeed; 2/ the pkg and  struct that handles
comparison logic; and 3/  the set of scopes that
compare values in details.ItemInfo instead of in
the item's storage or logical path.  The third
category is the primary overload here, and is
better off with a different name.