observe resourceOwner once, at top
we transitioned backups from multi-user to single-
user some time ago, but never went back to
remove the resource-owner identification from
every progress bar on display.  This ends up
creating a lot of duplication of the same owner
in a single run, when we could use that space
for more valuable information such as item
or folder details.