Create an abstraction for location representations
Create an abstraction that can represent a unique location but also turn
that into a location that is stored in backup details. This removes the
need for other components to know what the precise transform is between
the two while still allowing the prefix matcher to use something with
less strict rules than path.Path.