Create way to recover unique loc from details
Previous LocationRefs may be sourced from either the LocationRef field
of the details item or the RepoRef if it is an old backup for OneDrive
or SharePoint. The code to source from RepoRef can be removed after a
while (see comments in code).

Sourcing transform values from the prefix matcher can also be simplified
after some time as right now it assumes the LocationRef may not be
available for Exchange in some cases. After a sufficiently long time
(similar to requirements for OneDrive/SharePoint above), the first
prefix matcher lookup can be removed and we can just always use the
parsed previous LocationRef.