Don't mark items updated while migrating
When first populating LocationRefs, try to get the updated flag right be
tuning the "updated" checks a little. Specifically, don't force an item
to be marked as updated if the previous LocationRef for the item was
empty. Comparisons on the RepoRef for the item should catch any actual
updates during this period.