specify backup version in backup operation (#3170)
In order to test version-based changes at the operation level, we need
to control what version is persisted in the backup. This has some edge
cases, for instance: setting the backup version doesn't actually force
behavior to comply with that version's behavior. It only writes that
version to the backup model. For that reason, new backup models default
to the current version, and must be manually overwritten if a test needs
a different variation.