Struct to hold path information for restore
Make a struct to hold path information for restore operations. The
struct contains the location the item is located at in storage (kopia)
and the collection the item should be placed into and/or restored to
(RestorePath).

Placed in the path package because this will need to be imported by many
disjoint packages including onedrive, operations, and kopia.