minor updates to the path.Path interface
Swaps path.Append from a single item method to accept
a variadic list of string elements.  Since 95% of all calls
to path.Append were items, also adds a shorthand
AppendItem func to the interface for easy clarity.
Finally, adds a Last() method to elements for getting
the last element in the slice.