Make sure to close the body
From https://pkg.go.dev/net/http#Client.Do

> If the returned error is nil, the Response will contain a non-nil
> Body which the user is expected to close. If the Body is not both read
> to EOF and closed, the Client's underlying RoundTripper (typically
> Transport) may not be able to re-use a persistent TCP connection to
> the server for a subsequent "keep-alive" request.

These were found using https://github.com/timakin/bodyclose. Not
adding it as a linter for now as it was producing false positive(I
think) on OneDrive item. The close was happening in
https://github.com/alcionai/corso/blob/a9918d2f78b9419af553a6e0356bbf5f6078b9d8/src/internal/kopia/upload.go#L68
instead of right alongside it which might be why it got thrown off.