Unexport things, wrap for compatibility
Make the new FindBases function return the same type as the older
fucntion so we can use it without changing the caller. It will be
inefficient as we'll do some lookups multiple times, but allows us to
wrap this up quickly and come back to make things nicer later on.