auto-log recoverable errors with stack
automatically log when we add a recoverable error or a skipped
item to fault.  This log will include a stack trace of the call from
the location of the logged recoverable.  Clues does not have a
method for pulling a stack trace out of an error yet; that can be
added at a future date.