final cleanup for renaming/movement
This last step in the cleanup has two goals: first to minimize the
number of arbitrarily named and located files so that code is better
condensed and co-located.  This is all just file renaming and a
minor amount of code copy-pasting.

The second is to create a centralized package to own the ColInfo
type structs that we use to both stub out test data, and also generate
factory data for cmds.  The current ownership is haphazard, and while
this movement is a little more condensed, it's still jumping through
some weird hoops to get things to work.  Treat it as one good step
forward, and we'll have to return to polish it another time.  At least
it'll be separated from the m365 folder at large this way, and more
easily identified as supporting design rather than production usage.