move default rate limit to per-second
The current rate limit default maximizes the number of
calls we can make in the first 10 minutes, but the larger
initial bucket size means we have to drop the per-second
token burst.  As a result, any process longer than 10 minutes
is going to see unnecessary rate limiting by ~100 calls
per minute.

This change swaps the large initial bucket for a smaller
one, and a larger token burst to increase throughput
across long runtimes.