handle file-folder collision cases
handle onedrive item collision cases where a file
to be restored is in conflict with an existing folder.
Skip and Copy will function normally.  For Replace
behavior, we'll defer to Copy in this situation, so that
the file gets restored without changing or deleting
the folder.  If a folder creation attempts to replace
an item, we do a similar action and make a renamed
copy of the folder instead.