fix bad user info conditional on spo license
Primary change is fixing the spo license check in
GetUserInfo from a || to a && !.  Currently the check
returns the spo license error instead of treating it
as a drive access condition.

Additionally adds a root site check in the m365
isServiceEnabled call, instead of always returning
true for sharepoint.  This should catch spo license
issues on a per-site basis, if someone makes it that
far.

Finally, adds extra testing (plus some test code
consolidation and refactoring) to ensure GetUserInfo
is properly tested.