Add e2e tests for boomeranging mail and files
Add a test that will fail if something is wrong with tracking items that
are moved from a folder to a different folder and back to the original
folder between incremental backups. This is important as it will help us
ensure we don't accidentally turn on Exchange immutable IDs before
they're fixed.