Refactor/expand various kopia backup tests
Mostly basic tests, but check various situations when it comes to
incremental backups with respect to hierarchy merging and kopia-assisted
incrementals. This helps build confidence that even in more complex
situations building the tree in a certain way will work as expected. It
also shows that it's possible and still efficient to run backups without
kopia-assisted incremental bases.