Add 'halves()' to path interface
Adds a func to the path interface which breaks up the path
into its two distinct halves: the prefix builder, and the folder
and item elements.  Likely to be useful in the future where
prefixes will be of arbitrary length, and downstream consumers
are primarly focused on the non-prefix portion of the path.