remove all uses of iota
I've needed to catch gotchas that arise from contributors
adding a value in the middle of an iota list, not to mention
have dealt with prior bugs that happened the same way,
now too many times to feel safe about its usage.

This PR removes the use of iota from all const declarations.
The intent is to not allow the use of iota within the codebase.