add non-delta message paging path
similar to exchange, there are conditions under
which channel messages are unable to use the
delta api.  In particular, when the channel has no
valid email property.  This change ensures that
we enumerate the entire channel under those
conditions, instead of failing out.

Technically, the only situation we know of where
the email property is missing, and thus the channel
cannot make delta queries, is when there are
zero messages in the channel already.  However,
there may be unknown cases that we haven't caught,
so instead of skipping the channel we're going to
handle non-delta enumeration for future safeguarding.